+++ packages/dds/tree/src/test/simple-tree/types.spec.ts
@@ -17,7 +17,7 @@ import { RawTreeNode } from "../../simple-tree/rawNode.js";
 import { numberSchema } from "../../simple-tree/leafNodeSchema.js";
 // eslint-disable-next-line import/no-internal-modules
 import { getFlexSchema } from "../../simple-tree/toFlexSchema.js";
-import { validateUsageError } from "./utils.js";
+import { validateUsageError } from "../utils.js";
 
 describe("simple-tree types", () => {
 	describe("TreeNode", () => {
